backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.15.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
PyJWT==2.10.1
passlib==1.7.4
python-multipart==0.0.20
redis==5.2.1
httpx==0.28.1
pytest==8.3.5
